Renamed packages in sleepiq and sensebox bindings
