+++ addons/binding/org.openhab.binding.sensebox/src/main/java/org/openhab/binding/sensebox/internal/SenseBoxAPIConnection.java
@@ -12,11 +12,11 @@ import com.google.gson.Gson;
 import org.apache.commons.lang.StringUtils;
 import org.eclipse.smarthome.core.thing.ThingStatus;
 import org.eclipse.smarthome.io.net.http.HttpUtil;
-import org.openhab.binding.sensebox.model.SenseBoxData;
-import org.openhab.binding.sensebox.model.SenseBoxDescriptor;
-import org.openhab.binding.sensebox.model.SenseBoxLoc;
-import org.openhab.binding.sensebox.model.SenseBoxLocation;
-import org.openhab.binding.sensebox.model.SenseBoxSensor;
+import org.openhab.binding.sensebox.internal.model.SenseBoxData;
+import org.openhab.binding.sensebox.internal.model.SenseBoxDescriptor;
+import org.openhab.binding.sensebox.internal.model.SenseBoxLoc;
+import org.openhab.binding.sensebox.internal.model.SenseBoxLocation;
+import org.openhab.binding.sensebox.internal.model.SenseBoxSensor;
 import org.osgi.framework.FrameworkUtil;
 import org.osgi.framework.Version;
 import org.slf4j.Logger;
