+++ addons/binding/org.openhab.binding.sensebox/src/main/java/org/openhab/binding/sensebox/internal/model/SenseBoxLocation.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.sensebox.model;
+package org.openhab.binding.sensebox.internal.model;
 
 /**
  * The {@link SenseBoxLocation} holds a de-serialized representation
