+++ addons/binding/org.openhab.binding.sleepiq/src/main/java/org/openhab/binding/sleepiq/internal/SleepIQHandlerFactory.java
@@ -20,9 +20,9 @@ import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.ThingUID;
 import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
-import org.openhab.binding.sleepiq.discovery.SleepIQBedDiscoveryService;
 import org.openhab.binding.sleepiq.handler.SleepIQCloudHandler;
 import org.openhab.binding.sleepiq.handler.SleepIQDualBedHandler;
+import org.openhab.binding.sleepiq.internal.discovery.SleepIQBedDiscoveryService;
 import org.osgi.framework.ServiceRegistration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
